Split lazy chunk error classification
