+++ src/renderer/src/lib/lazy-with-retry.ts
@@ -1,4 +1,9 @@
 import { lazy, type ComponentType, type LazyExoticComponent } from 'react'
+import {
+  describeLazyChunkError,
+  isAppRestartEligibleLazyChunkError,
+  type LazyChunkErrorBreadcrumbData
+} from './lazy-chunk-error-classification'
 
 /**
  * Resilient replacement for React.lazy.
@@ -55,13 +60,6 @@ type ReloadGuardStatus =
   | 'reload-requested'
   | 'not-attempted'
   | 'unreadable'
-
-type LazyChunkErrorBreadcrumbData = {
-  reloadKey: string
-  errorName: string
-  errorCategory: string
-  messageClass: string
-}
 
 function readChunkReloadGuard(): ReloadGuardStatus {
   try {
@@ -233,92 +231,6 @@ async function restartAppForChunkRecovery(data: LazyChunkErrorBreadcrumbData): P
   const restarted = await restartAttempt
   appRestartInFlightByVersion.delete(appVersion)
   return restarted
-}
-
-function describeLazyChunkError(reloadKey: string, error: unknown): LazyChunkErrorBreadcrumbData {
-  const errorName = classifyErrorName(error)
-  const message = error instanceof Error ? error.message : stringifyUnknown(error)
-  const messageClass = classifyErrorMessage(message)
-  return {
-    reloadKey: sanitizeReloadKey(reloadKey),
-    errorName,
-    errorCategory: classifyErrorCategory(messageClass),
-    messageClass
-  }
-}
-
-function classifyErrorName(error: unknown): string {
-  if (!(error instanceof Error)) {
-    return 'NonError'
-  }
-  switch (error.name) {
-    case 'AggregateError':
-    case 'DOMException':
-    case 'EvalError':
-    case 'Error':
-    case 'RangeError':
-    case 'ReferenceError':
-    case 'SyntaxError':
-    case 'TypeError':
-    case 'URIError':
-      return error.name
-    default:
-      return 'Error'
-  }
-}
-
-function classifyErrorMessage(message: string): string {
-  const normalized = message.toLowerCase()
-  const looksLikeJsonParseFailure =
-    normalized.includes('json') || normalized.includes('not valid json')
-  if (
-    (!looksLikeJsonParseFailure &&
-      (normalized.includes('unexpected token') ||
-        normalized.includes('unexpected end of input'))) ||
-    normalized.includes('illegal return') ||
-    normalized.includes('import declarations may only appear') ||
-    normalized.includes('missing ) after argument list')
-  ) {
-    return 'syntax'
-  }
-  if (
-    normalized.includes('failed to fetch dynamically imported module') ||
-    normalized.includes('error loading dynamically imported module') ||
-    normalized.includes('importing a module script failed') ||
-    normalized.includes('loading chunk') ||
-    normalized.includes('chunkloaderror') ||
-    normalized.includes('networkerror')
-  ) {
-    return 'fetch'
-  }
-  return 'unknown'
-}
-
-function classifyErrorCategory(messageClass: string): string {
-  if (messageClass === 'syntax') {
-    return 'syntax'
-  }
-  if (messageClass === 'fetch') {
-    return 'fetch'
-  }
-  return 'unknown'
-}
-
-function isAppRestartEligibleLazyChunkError(data: LazyChunkErrorBreadcrumbData): boolean {
-  return data.errorCategory === 'syntax' || data.errorCategory === 'fetch'
-}
-
-function sanitizeReloadKey(reloadKey: string): string {
-  const trimmed = reloadKey.trim()
-  return /^[a-z0-9._:-]{1,80}$/i.test(trimmed) ? trimmed : 'unknown'
-}
-
-function stringifyUnknown(value: unknown): string {
-  try {
-    return String(value)
-  } catch {
-    return ''
-  }
 }
 
 const wait = (ms: number): Promise<void> => new Promise((resolve) => setTimeout(resolve, ms))
